+++ compiler-rt/lib/scudo/standalone/tests/release_test.cpp
@@ -138,18 +138,15 @@ TEST(ScudoReleaseTest, FreePagesRangeTracker) {
 
 class ReleasedPagesRecorder {
 public:
-  ReleasedPagesRecorder() = default;
-  explicit ReleasedPagesRecorder(scudo::uptr Base) : Base(Base) {}
   std::set<scudo::uptr> ReportedPages;
 
   void releasePageRangeToOS(scudo::uptr From, scudo::uptr To) {
     const scudo::uptr PageSize = scudo::getPageSizeCached();
     for (scudo::uptr I = From; I < To; I += PageSize)
-      ReportedPages.insert(I + getBase());
+      ReportedPages.insert(I);
   }
 
-  scudo::uptr getBase() const { return Base; }
-  scudo::uptr Base = 0;
+  scudo::uptr getBase() const { return 0; }
 };
 
 // Simplified version of a TransferBatch.
@@ -222,8 +219,7 @@ template <class SizeClassMap> void testReleaseFreeMemoryToOS() {
     ReleasedPagesRecorder Recorder;
     scudo::PageReleaseContext Context(BlockSize,
                                       /*RegionSize=*/MaxBlocks * BlockSize,
-                                      /*NumberOfRegions=*/1U,
-                                      /*ReleaseSize=*/MaxBlocks * BlockSize);
+                                      /*NumberOfRegions=*/1U);
     ASSERT_FALSE(Context.hasBlockMarked());
     Context.markFreeBlocks(FreeList, DecompactPtr, Recorder.getBase());
     ASSERT_TRUE(Context.hasBlockMarked());
@@ -329,9 +325,8 @@ template <class SizeClassMap> void testPageMapMarkRange() {
       const scudo::uptr GroupEnd = GroupBeg + GroupSize;
 
       scudo::PageReleaseContext Context(BlockSize, RegionSize,
-                                        /*NumberOfRegions=*/1U,
-                                        /*ReleaseSize=*/RegionSize);
-      Context.markRangeAsAllCounted(GroupBeg, GroupEnd, /*Base=*/0U);
+                                        /*NumberOfRegions=*/1U);
+      Context.markRangeAsAllCounted(GroupBeg, GroupEnd, /*Base=*/0);
 
       scudo::uptr FirstBlock =
           ((GroupBeg + BlockSize - 1) / BlockSize) * BlockSize;
@@ -399,139 +394,10 @@ template <class SizeClassMap> void testPageMapMarkRange() {
 
     // Release the entire region. This is to ensure the last page is counted.
     scudo::PageReleaseContext Context(BlockSize, RegionSize,
-                                      /*NumberOfRegions=*/1U,
-                                      /*ReleaseSize=*/RegionSize);
+                                      /*NumberOfRegions=*/1U);
     Context.markRangeAsAllCounted(/*From=*/0U, /*To=*/RegionSize, /*Base=*/0);
     for (scudo::uptr Page = 0; Page < RoundedRegionSize / PageSize; ++Page)
       EXPECT_TRUE(Context.PageMap.isAllCounted(/*Region=*/0, Page));
-  } // Iterate each size class
-}
-
-template <class SizeClassMap> void testReleasePartialRegion() {
-  typedef FreeBatch<SizeClassMap> Batch;
-  const scudo::uptr PageSize = scudo::getPageSizeCached();
-  const scudo::uptr ReleaseBase = PageSize;
-  const scudo::uptr BasePageOffset = ReleaseBase / PageSize;
-
-  for (scudo::uptr I = 1; I <= SizeClassMap::LargestClassId; I++) {
-    // In the following, we want to ensure the region includes at least 2 pages
-    // and we will release all the pages except the first one. The handling of
-    // the last block is tricky, so we always test the case that includes the
-    // last block.
-    const scudo::uptr BlockSize = SizeClassMap::getSizeByClassId(I);
-    const scudo::uptr RegionSize =
-        scudo::roundUpSlow(scudo::roundUp(BlockSize, PageSize) * 2, BlockSize) +
-        BlockSize;
-    const scudo::uptr RoundedRegionSize = scudo::roundUp(RegionSize, PageSize);
-
-    scudo::SinglyLinkedList<Batch> FreeList;
-    FreeList.clear();
-
-    // Skip the blocks in the first page and add the remaining.
-    std::vector<scudo::uptr> Pages(RoundedRegionSize / PageSize, 0);
-    for (scudo::uptr Block = scudo::roundUpSlow(PageSize, BlockSize);
-         Block + BlockSize <= RoundedRegionSize; Block += BlockSize) {
-      for (scudo::uptr Page = Block / PageSize;
-           Page <= (Block + BlockSize - 1) / PageSize; ++Page) {
-        ASSERT_LT(Page, Pages.size());
-        ++Pages[Page];
-      }
-    }
-
-    // This follows the logic how we count the last page. It should be
-    // consistent with how markFreeBlocks() handles the last block.
-    if (RoundedRegionSize % BlockSize != 0)
-      ++Pages.back();
-
-    Batch *CurrentBatch = nullptr;
-    for (scudo::uptr Block = scudo::roundUpSlow(PageSize, BlockSize);
-         Block < RegionSize; Block += BlockSize) {
-      if (CurrentBatch == nullptr ||
-          CurrentBatch->getCount() == Batch::MaxCount) {
-        CurrentBatch = new Batch;
-        CurrentBatch->clear();
-        FreeList.push_back(CurrentBatch);
-      }
-      CurrentBatch->add(Block);
-    }
-
-    auto VerifyReleaseToOs = [&](scudo::PageReleaseContext &Context) {
-      auto SkipRegion = [](UNUSED scudo::uptr RegionIndex) { return false; };
-      ReleasedPagesRecorder Recorder(ReleaseBase);
-      releaseFreeMemoryToOS(Context, Recorder, SkipRegion);
-      const scudo::uptr FirstBlock = scudo::roundUpSlow(PageSize, BlockSize);
-
-      for (scudo::uptr P = 0; P < RoundedRegionSize; P += PageSize) {
-        if (P < FirstBlock) {
-          // If FirstBlock is not aligned with page boundary, the first touched
-          // page will not be released either.
-          EXPECT_TRUE(Recorder.ReportedPages.find(P) ==
-                      Recorder.ReportedPages.end());
-        } else {
-          EXPECT_TRUE(Recorder.ReportedPages.find(P) !=
-                      Recorder.ReportedPages.end());
-        }
-      }
-    };
-
-    // Test marking by visiting each block.
-    {
-      auto DecompactPtr = [](scudo::uptr P) { return P; };
-      scudo::PageReleaseContext Context(
-          BlockSize, RegionSize, /*NumberOfRegions=*/1U,
-          /*ReleaseSize=*/RegionSize - PageSize, ReleaseBase);
-      Context.markFreeBlocks(FreeList, DecompactPtr, /*Base=*/0U);
-      for (const Batch &It : FreeList) {
-        for (scudo::u16 I = 0; I < It.getCount(); I++) {
-          scudo::uptr Block = It.get(I);
-          for (scudo::uptr Page = Block / PageSize;
-               Page <= (Block + BlockSize - 1) / PageSize; ++Page) {
-            EXPECT_EQ(Pages[Page], Context.PageMap.get(/*Region=*/0U,
-                                                       Page - BasePageOffset));
-          }
-        }
-      }
-
-      VerifyReleaseToOs(Context);
-    }
-
-    // Test range marking.
-    {
-      scudo::PageReleaseContext Context(
-          BlockSize, RegionSize, /*NumberOfRegions=*/1U,
-          /*ReleaseSize=*/RegionSize - PageSize, ReleaseBase);
-      Context.markRangeAsAllCounted(ReleaseBase, RegionSize, /*Base=*/0U);
-      for (scudo::uptr Page = ReleaseBase / PageSize;
-           Page < RoundedRegionSize / PageSize; ++Page) {
-        if (Context.PageMap.get(/*Region=*/0, Page - BasePageOffset) !=
-            Pages[Page]) {
-          EXPECT_TRUE(Context.PageMap.isAllCounted(/*Region=*/0,
-                                                   Page - BasePageOffset));
-        }
-      }
-
-      VerifyReleaseToOs(Context);
-    }
-
-    // Check the buffer size of PageMap.
-    {
-      scudo::PageReleaseContext Full(BlockSize, RegionSize,
-                                     /*NumberOfRegions=*/1U,
-                                     /*ReleaseSize=*/RegionSize);
-      Full.ensurePageMapAllocated();
-      scudo::PageReleaseContext Partial(
-          BlockSize, RegionSize, /*NumberOfRegions=*/1U,
-          /*ReleaseSize=*/RegionSize - PageSize, ReleaseBase);
-      Partial.ensurePageMapAllocated();
-
-      EXPECT_GE(Full.PageMap.getBufferSize(), Partial.PageMap.getBufferSize());
-    }
-
-    while (!FreeList.empty()) {
-      CurrentBatch = FreeList.front();
-      FreeList.pop_front();
-      delete CurrentBatch;
-    }
   } // Iterate each size class
 }
 
@@ -552,11 +418,4 @@ TEST(ScudoReleaseTest, PageMapMarkRange) {
   testPageMapMarkRange<scudo::AndroidSizeClassMap>();
   testPageMapMarkRange<scudo::FuchsiaSizeClassMap>();
   testPageMapMarkRange<scudo::SvelteSizeClassMap>();
-}
-
-TEST(ScudoReleaseTest, ReleasePartialRegion) {
-  testReleasePartialRegion<scudo::DefaultSizeClassMap>();
-  testReleasePartialRegion<scudo::AndroidSizeClassMap>();
-  testReleasePartialRegion<scudo::FuchsiaSizeClassMap>();
-  testReleasePartialRegion<scudo::SvelteSizeClassMap>();
 }