+++ libc/src/unistd/linux/fork.cpp
@@ -13,7 +13,7 @@
 #include "src/__support/threads/fork_callbacks.h"
 #include "src/__support/threads/thread.h" // For thread self object
 
-#include <src/errno/libc_errno.h>
+#include "src/errno/libc_errno.h"
 #include <sys/syscall.h> // For syscall numbers.
 
 namespace __llvm_libc {