+++ libc/src/unistd/linux/linkat.cpp
@@ -11,8 +11,8 @@
 #include "src/__support/OSUtil/syscall.h" // For internal syscall function.
 #include "src/__support/common.h"
 
+#include "src/errno/libc_errno.h"
 #include <fcntl.h>
-#include <src/errno/libc_errno.h>
 #include <sys/syscall.h> // For syscall numbers.
 
 namespace __llvm_libc {