+++ mlir/include/mlir/Dialect/Linalg/Transforms/Transforms.h
@@ -531,10 +531,10 @@ FailureOr<LinalgOp> promoteSubViews(OpBuilder &b, LinalgOp op,
                                     const LinalgPromotionOptions &options);
 
 /// Allocate the subview in the GPU workgroup memory.
-Optional<Value> allocateWorkgroupMemory(OpBuilder &builder,
-                                        memref::SubViewOp subview,
-                                        ArrayRef<Value> sizeBounds,
-                                        DataLayout &);
+std::optional<Value> allocateWorkgroupMemory(OpBuilder &builder,
+                                             memref::SubViewOp subview,
+                                             ArrayRef<Value> sizeBounds,
+                                             DataLayout &);
 
 /// In case of GPU group memory there is no need to deallocate.
 LogicalResult deallocateWorkgroupMemory(OpBuilder &, Value /*buffer*/);
@@ -544,10 +544,10 @@ LogicalResult deallocateWorkgroupMemory(OpBuilder &, Value /*buffer*/);
 LogicalResult copyToWorkgroupMemory(OpBuilder &b, Value src, Value dst);
 
 /// Allocate the subview in the GPU private memory.
-Optional<Value> allocateGPUPrivateMemory(OpBuilder &builder,
-                                         memref::SubViewOp subview,
-                                         ArrayRef<Value> sizeBounds,
-                                         DataLayout &);
+std::optional<Value> allocateGPUPrivateMemory(OpBuilder &builder,
+                                              memref::SubViewOp subview,
+                                              ArrayRef<Value> sizeBounds,
+                                              DataLayout &);
 
 /// Normal copy to between src and dst.
 LogicalResult copyToGPUPrivateMemory(OpBuilder &b, Value src, Value dst);